requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
imageio==2.34.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
scikit-image==0.22.0
scikit-learn==1.0.2
scipy==1.13.0
threadpoolctl==3.4.0
tifffile==2024.2.12
Werkzeug==3.0.2
zipp==3.18.1
